Always store as float32

+++ geoh5py/data/numeric_data.py
@@ -81,7 +81,7 @@ class NumericData(Data, ABC):
             values = np.ravel(values)
             warn("Input 'values' converted to a 1D array.")
 
-        values = values.astype(float)
+        values = values.astype(np.float32)
 
         # change nan values to nan_value
         values[np.isnan(values)] = self.nan_value
